+++ packages/create-purplet/build.js
@@ -2,7 +2,6 @@ import { rollup } from 'rollup';
 import { copyFile, mkdir, readdir, stat } from 'fs/promises';
 import { existsSync } from 'fs';
 import path from 'path';
-import minimatch from 'minimatch';
 
 process.env.NODE_ENV = 'production';
 
@@ -13,7 +12,7 @@ await roll.write(rollupConfig.output);
 
 const examples = (await readdir('../../examples')).filter(x =>
   existsSync(path.join('../../examples', x, '.template.json'))
-);
+).filter(x => !x.startsWith('local-'))
 
 async function mkdirp(root) {
   try {