+++ packages/create-purplet/CHANGELOG.md
@@ -1,5 +1,11 @@
 # `create-purplet`
 
+## 2.0.2
+
+- Fix issues running with `yarn` and `npm`
+- Fix issues running on windows at all
+- For simplicity, it does not install packages automatically anymore
+
 ## 2.0.1
 
 - Fix issues locating the templates folder.